minor refactor to adhere to cosmos
